move Repo.diff related test to diff
